ARM: dts: imx: add support for the ATM0700D4 panel attached to sk-imx53
The SK-ATM0700D4-Plug is an extension board (provided by the same
manufacturer, [1]) which can be connected to the SK-IMX53 panel kit. The
panel can be connected either using the RGB parallel bus or using the
LVDS connector (recommended). Add DT files describing this "shield",
both RGB and LVDS connections.

[1] http://starterkit.ru/html/index.php?name=shop&op=view&id=64

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 8c17cec6347d9e0ab2c6005bd1804067e12ca4e0 ]