ARM: dts: renesas: marzen: Add FLASH node
Add a device node for the Spansion S29GL512N NOR FLASH on the Marzen
development board.  This FLASH resides in the external address space of
the Local Bus State Controller.

Note that as the CFI-FLASH has a run-time conflict with CPU bring-up, it
will only be available when booting with SMP disabled.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Link: https://lore.kernel.org/r/440ce3bb950c34fd57071b4eec83ad9643b682e5.1693481518.git.geert+renesas@glider.be

[ upstream commit: 6ebf1725e1f0bf712f547de7d9d14066279bb546 ]