dt-bindings: arm: rockchip: Update edgeble-neu6 bindings
Current binding has following compatible combination for SoM, IO board
- NCM6A: edgeble,neural-compute-module-6a,
	 edgeble,neural-compute-module-6a-io
- NCM6B: edgeble,neural-compute-module-6b,
	 edgeble,neural-compute-module-6b-io

This board topology now changes in final revisions, so a common
compatible IO board, NCM6A-IO, is made available for both variants
of SoM: NCM6A and NCM6B, produced by Edgeble AI.

With this change, 6b-io is not as available and 6a; 6b SoM's have the
same compatible IO board as 6a-io. This change is due to the common
optimised design of the IO board made available now in final revisions,
which was not the case in initial revisions.

Update the dt-bindings with this new change.

Signed-off-by: Jagan Teki <jagan@edgeble.ai>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20231125142914.57459-1-jagan@edgeble.ai
Signed-off-by: Heiko Stuebner <heiko@sntech.de>

[ upstream commit: b685460632d51d7307bbba3d173c0dd6dbf38fa8 ]