media: dt-bindings: samsung,s5p-mfc: Fix iommu properties schemas
The iommus and iommu-names property schemas have several issues. First,
'iommus-names' in the if/then schemas is the wrong name. As all the names
are the same, they can be defined at the top level instead. Then the
if/then schemas just need to define how many entries. The iommus if/then
schemas are also redundant. Best I can tell, the desire was to require 2
entries for "samsung,exynos5433-mfc", "samsung,mfc-v5", "samsung,mfc-v6",
and "samsung,mfc-v8".

Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Reviewed-by: Aakarsh Jain <aakarsh.jain@samsung.com>
Link: https://lore.kernel.org/r/20231214195553.862920-1-robh@kernel.org
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: a1499b7541cc26bac360cf07e49fd1e2ab3dd17f ]