Revert "arm64: dts: imx8mn-var-som-symphony: Describe the USB-C connector"
This reverts commit 095b96b2b8c6dcabf40bbfe4bb99a95fe55c7463.

Marek reported the similar change in imx8mp-dhcom-pdk3.dts
caused a regression:

"This patch breaks USB-C port on this board.

If I plug in USB-C storage device, it is not detected. If I revert this
patch, it is detected.

Please drop this patch for now."

Revert it here as well.

Fixes: 095b96b2b8c6 ("arm64: dts: imx8mn-var-som-symphony: Describe the USB-C connector")
Signed-off-by: Fabio Estevam <festevam@denx.de>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 690085d866f08cc72ae4d601821564a0b63e32f3 ]