dt-bindings: drm/bridge: ti-sn65dsi86: Fix bouncing @codeaurora address
The servers for the @codeaurora domain are long retired and any messages
sent there bounce.  Sandeep Panda's email address is no longer valid and
should be repleaced.  However Sandeep has left the company and has not
been active sice, therefore it looks like this binding is orphaned.

Doug is listed as the reviewer for this file in MAINTAINERS and has
volunteered to be listed within the file as the binding maintainer.
Therefore replace Sandeep with Doug to make the documentation current.

Signed-off-by: Jeffrey Hugo <quic_jhugo@quicinc.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Reviewed-by: Neil Armstrong <neil.armstrong@linaro.org>
Acked-by: Rob Herring <robh@kernel.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20240202202329.4172917-1-quic_jhugo@quicinc.com

[ upstream commit: 2592a36d956955f0e0f399d608d7be0ec280a40c ]