ARM: dts: qcom: samsung-matisse-common: Add initial common device tree
According to the dts from the kernel source code released by Samsung,
matissewifi and matisselte only have minor differences in hardware, so
use a shared dtsi to reduce duplicated code. Additionally, this should
make adding support for matisse3g easier should someone want to do that
at a later point.

As such, add a common device tree for all matisse devices by Samsung
based on the matissewifi dts. Support for matisselte will be introduced
in a later patch in this series and will use the common dtsi as well.

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20240215180322.99089-2-newbyte@postmarketos.org
Signed-off-by: Bjorn Andersson <andersson@kernel.org>

[ upstream commit: f91dc3e6f3840417bd0456cd46b5bd8cea288f31 ]