regulator: dt-bindings: gpio-regulator: Fix "gpios-states" and "states" array bounds
The minimum number of array entries for "gpios-states" was not not
specified, so the the default is the same as the max (8).

The minimum is also missing from "states", and the maximum is also wrong
as it should be 2^(# of GPIO lines). Since there can be 1 to 8 GPIOs,
the "states" range should be 2 to 256.

Signed-off-by: Rob Herring <robh@kernel.org>
Link: https://msgid.link/r/20240224000752.3830665-1-robh@kernel.org
Signed-off-by: Mark Brown <broonie@kernel.org>

[ upstream commit: efe9a6da2292e68d45cd9e0eb1c31b85841d0c2a ]